Extract aggregate business logic to ontology package
