+++ concept/service.go
@@ -24,10 +24,8 @@ import (
 )
 
 const (
-	smartlogicAuthority      = "Smartlogic"
-	managedLocationAuthority = "ManagedLocation"
-	thingsAPIEndpoint        = "/things"
-	conceptsAPIEnpoint       = "/concepts"
+	thingsAPIEndpoint  = "/things"
+	conceptsAPIEnpoint = "/concepts"
 )
 
 var irregularConceptTypePaths = map[string]string{
@@ -311,19 +309,19 @@ func bucketConcordances(concordanceRecords []concordances.ConcordanceRecord) (ma
 
 	var primaryAuthority string
 	var err error
-	slRecords, slFound := bucketedConcordances[smartlogicAuthority]
+	slRecords, slFound := bucketedConcordances[ontology.SmartlogicAuthority]
 	if slFound {
 		if len(slRecords) == 1 {
-			primaryAuthority = smartlogicAuthority
+			primaryAuthority = ontology.SmartlogicAuthority
 		} else {
 			err = fmt.Errorf("more than 1 primary authority")
 		}
 	}
-	mlRecords, mlFound := bucketedConcordances[managedLocationAuthority]
+	mlRecords, mlFound := bucketedConcordances[ontology.ManagedLocationAuthority]
 	if mlFound {
 		if len(mlRecords) == 1 {
 			if primaryAuthority == "" {
-				primaryAuthority = managedLocationAuthority
+				primaryAuthority = ontology.ManagedLocationAuthority
 			}
 		} else {
 			err = fmt.Errorf("more than 1 ManagedLocation primary authority")
@@ -417,41 +415,9 @@ func (s *AggregateService) getConcordedConcept(ctx context.Context, UUID string,
 		sources = append(sources, primaryConcept)
 	}
 
-	var scopeNoteOptions = map[string][]string{}
-	concordedConcept := ontology.ConcordedConcept{}
-	for _, src := range sources {
-		concordedConcept = mergeCanonicalInformation(concordedConcept, src, scopeNoteOptions)
-	}
-
-	concordedConcept.Aliases = deduplicateAndSkipEmptyAliases(concordedConcept.Aliases)
-	concordedConcept.ScopeNote = chooseScopeNote(concordedConcept, scopeNoteOptions)
+	concordedConcept := ontology.CreateAggregateConcept(sources)
 
 	return concordedConcept, transactionID, nil
-}
-
-func chooseScopeNote(concept ontology.ConcordedConcept, scopeNoteOptions map[string][]string) string {
-	if sn, ok := scopeNoteOptions[smartlogicAuthority]; ok {
-		return strings.Join(removeMatchingEntries(sn, concept.PrefLabel), " | ")
-	}
-	if sn, ok := scopeNoteOptions["Wikidata"]; ok {
-		return strings.Join(removeMatchingEntries(sn, concept.PrefLabel), " | ")
-	}
-	if sn, ok := scopeNoteOptions["TME"]; ok {
-		if concept.Type == "Location" {
-			return strings.Join(removeMatchingEntries(sn, concept.PrefLabel), " | ")
-		}
-	}
-	return ""
-}
-
-func removeMatchingEntries(slice []string, matcher string) []string {
-	var newSlice []string
-	for _, k := range slice {
-		if k != matcher {
-			newSlice = append(newSlice, k)
-		}
-	}
-	return newSlice
 }
 
 func (s *AggregateService) Healthchecks() []fthealth.Check {
@@ -467,171 +433,6 @@ func (s *AggregateService) Healthchecks() []fthealth.Check {
 		checks = append(checks, s.kinesis.Healthcheck())
 	}
 	return checks
-}
-
-func deduplicateAndSkipEmptyAliases(aliases []string) []string {
-	aMap := map[string]bool{}
-	var outAliases []string
-	for _, v := range aliases {
-		if v == "" {
-			continue
-		}
-		aMap[v] = true
-	}
-	for a := range aMap {
-		outAliases = append(outAliases, a)
-	}
-	return outAliases
-}
-
-func getMoreSpecificType(existingType string, newType string) string {
-
-	// Thing type shouldn't wipe things.
-	if newType == "Thing" && existingType != "" {
-		return existingType
-	}
-
-	// If we've already called it a PublicCompany, keep that information.
-	if existingType == "PublicCompany" && (newType == "Organisation" || newType == "Company") {
-		return existingType
-	}
-	return newType
-}
-
-func buildScopeNoteOptions(scopeNotes map[string][]string, s ontology.Concept) {
-	var newScopeNote string
-	if s.Authority == "TME" {
-		newScopeNote = s.PrefLabel
-	} else {
-		newScopeNote = s.ScopeNote
-	}
-	if newScopeNote != "" {
-		scopeNotes[s.Authority] = append(scopeNotes[s.Authority], newScopeNote)
-	}
-}
-
-func mergeCanonicalInformation(c ontology.ConcordedConcept, s ontology.Concept, scopeNoteOptions map[string][]string) ontology.ConcordedConcept {
-	c.PrefUUID = s.UUID
-	c.PrefLabel = s.PrefLabel
-	c.Type = getMoreSpecificType(c.Type, s.Type)
-	c.Aliases = append(c.Aliases, s.Aliases...)
-	c.Aliases = append(c.Aliases, s.PrefLabel)
-	if s.Strapline != "" {
-		c.Strapline = s.Strapline
-	}
-	if s.DescriptionXML != "" {
-		c.DescriptionXML = s.DescriptionXML
-	}
-	if s.ImageURL != "" {
-		c.ImageURL = s.ImageURL
-	}
-	if s.EmailAddress != "" {
-		c.EmailAddress = s.EmailAddress
-	}
-	if s.FacebookPage != "" {
-		c.FacebookPage = s.FacebookPage
-	}
-	if s.TwitterHandle != "" {
-		c.TwitterHandle = s.TwitterHandle
-	}
-	buildScopeNoteOptions(scopeNoteOptions, s)
-	if s.ShortLabel != "" {
-		c.ShortLabel = s.ShortLabel
-	}
-	if len(s.SupersededByUUIDs) > 0 {
-		c.SupersededByUUIDs = s.SupersededByUUIDs
-	}
-	if len(s.ParentUUIDs) > 0 {
-		c.ParentUUIDs = s.ParentUUIDs
-	}
-	if len(s.BroaderUUIDs) > 0 {
-		c.BroaderUUIDs = s.BroaderUUIDs
-	}
-	if len(s.RelatedUUIDs) > 0 {
-		c.RelatedUUIDs = s.RelatedUUIDs
-	}
-	c.SourceRepresentations = append(c.SourceRepresentations, s)
-	if s.ProperName != "" {
-		c.ProperName = s.ProperName
-	}
-	if s.ShortName != "" {
-		c.ShortName = s.ShortName
-	}
-	if len(s.TradeNames) > 0 {
-		c.TradeNames = s.TradeNames
-	}
-	if len(s.FormerNames) > 0 {
-		c.FormerNames = s.FormerNames
-	}
-	if s.CountryCode != "" {
-		c.CountryCode = s.CountryCode
-	}
-	if s.CountryOfRisk != "" {
-		c.CountryOfRisk = s.CountryOfRisk
-	}
-	if s.CountryOfIncorporation != "" {
-		c.CountryOfIncorporation = s.CountryOfIncorporation
-	}
-	if s.CountryOfOperations != "" {
-		c.CountryOfOperations = s.CountryOfOperations
-	}
-	if s.PostalCode != "" {
-		c.PostalCode = s.PostalCode
-	}
-	if s.YearFounded > 0 {
-		c.YearFounded = s.YearFounded
-	}
-	if s.LeiCode != "" {
-		c.LeiCode = s.LeiCode
-	}
-	if s.BirthYear > 0 {
-		c.BirthYear = s.BirthYear
-	}
-	if s.Salutation != "" {
-		c.Salutation = s.Salutation
-	}
-	if s.ISO31661 != "" {
-		c.ISO31661 = s.ISO31661
-	}
-
-	for _, mr := range s.MembershipRoles {
-		c.MembershipRoles = append(c.MembershipRoles, ontology.MembershipRole{
-			RoleUUID:        mr.RoleUUID,
-			InceptionDate:   mr.InceptionDate,
-			TerminationDate: mr.TerminationDate,
-		})
-	}
-
-	for _, ic := range s.NAICSIndustryClassifications {
-		c.NAICSIndustryClassifications = append(c.NAICSIndustryClassifications, ontology.NAICSIndustryClassification{
-			UUID: ic.UUID,
-			Rank: ic.Rank,
-		})
-	}
-
-	if s.OrganisationUUID != "" {
-		c.OrganisationUUID = s.OrganisationUUID
-	}
-	if s.PersonUUID != "" {
-		c.PersonUUID = s.PersonUUID
-	}
-	if s.InceptionDate != "" {
-		c.InceptionDate = s.InceptionDate
-	}
-	if s.TerminationDate != "" {
-		c.TerminationDate = s.TerminationDate
-	}
-	if s.FigiCode != "" {
-		c.FigiCode = s.FigiCode
-	}
-	if s.IssuedBy != "" {
-		c.IssuedBy = s.IssuedBy
-	}
-	if s.IndustryIdentifier != "" {
-		c.IndustryIdentifier = s.IndustryIdentifier
-	}
-	c.IsDeprecated = s.IsDeprecated
-	return c
 }
 
 func sendToPurger(ctx context.Context, client httpClient, baseURL string, conceptUUIDs []string, conceptType string, conceptTypesWithPublicEndpoints []string, tid string) error {
@@ -843,7 +644,7 @@ func isTypeAllowedInElastic(concordedConcept ontology.ConcordedConcept) bool {
 	case "Membership":
 		for _, sr := range concordedConcept.SourceRepresentations {
 			//Allow smartlogic curated memberships through to elasticsearch as we will use them to discover authors
-			if sr.Authority == "Smartlogic" {
+			if sr.Authority == ontology.SmartlogicAuthority {
 				return true
 			}
 		}
